requirements.txt
selenium
pytest
pytest-cov
pytest-django